draft: ketcher editor subscription
image drop position captured

chore: show alias for atom R#

ketcher editor events

chore: show alias for atom R#

ketcher editor events

refac: atom location update with selected image without loop

refac: atom location update with selected image without loop

refac: atom location update with selected image without loop

chore: editor selection atttibute used for identification

chore: editor selection atttibute used for identification

chore: editor selection atttibute used for identification

chore: editor selection atttibute used for identification

fix: revert to loop approach for moving a connection

feat: alias-atom position on image resize

chore: template resize

implementation with stack FILO approach

feat: atom place with image on correct coordinates done

chore: multi template binded movement

feat: compatible with other atoms on canvas

feat: on atom move stick to template

fix: move temaplte to center after atom add

feat: mutation observers for cleanup and clear canvas added

chore: image delete not possible

chore: templates replace with Br -> A(R#)

+++ app/javascript/src/utilities/UserTemplatesHelpers.js
@@ -50,7 +50,7 @@ const updateUserTemplateDetails = async (oldValue, newValue) => {
 
 const onEventListen = async (event) => {
   let { newValue, oldValue } = event;
-  if (newValue.length && oldValue.length) {
+  if (newValue && oldValue && newValue.length && oldValue.length) {
     newValue = JSON.parse(newValue);
     oldValue = JSON.parse(oldValue);
     if (event.key === key) { // matching key && deleteAllowed
